Move the data operations and index write-back out of NoteStore.swift
